pyproject.toml
[tool.poetry]
name = "spacereeds-plotter"
version = "0.1.0"
description = ""
authors = ["SpaceReeds <spacereeds@gmail.com>", "Jakub Fraś <jakub.fras@alogliwice.polsl.pl>"]
license = "GPL-3.0"
readme = "README.md"
packages = [{ include = "plotter" }]

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.10.0"
pydantic = "^2.10.6"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
